requirements.txt
# HTTP and Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
scrapy>=2.11.0
aiohttp>=3.9.0
selenium>=4.15.0

# Social Media APIs
tweepy>=4.14.0
snscrape>=0.7.0

# YouTube
youtube-transcript-api>=0.6.1
pytube>=15.0.0
google-api-python-client>=2.108.0

# Podcast/RSS
feedparser>=6.0.10

# Document Processing
PyPDF2>=3.0.1
ebooklib>=0.18

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Vector Store
pinecone-client>=3.0.0
chromadb>=0.4.18
weaviate-client>=3.26.0

# AI/ML
openai>=1.6.0
sentence-transformers>=2.2.2
tiktoken>=0.5.2

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
click>=8.1.7
pandas>=2.1.4
tqdm>=4.66.1
tenacity>=8.2.3
ratelimit>=2.2.1
python-dateutil>=2.8.2

# Logging and Monitoring
loguru>=0.7.2
rich>=13.7.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
responses>=0.24.1

# Data Validation
pydantic>=2.5.3
validators>=0.22.0
